+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.22.20
 
 - Reduced minimum description length to 50 characters.
+- Filter SDK package `flutter` from the outdated report.
 
 ## 0.22.19